add default getter for decimals for existing pools
